[XLA:GPU] Propagate all profiling failures to `gemm_fusion_autotuner.cc`.
Currently resource exhausted failures are converted to `std::nullopt` which are then special cased (either ignored or converted back to internal failures). We remove the intermediate conversion and forward the failure to the caller. This is simpler and semantically equivalent with the added benefit that we don't loose the failure details.

PiperOrigin-RevId: 704685252