Fix an issue in `PartitionGatherTrivialSlicedOperandDimensions` when handling out-of-bound indices.
A gather operation will clamp the fetched indices such that we always retrieve the corresponding entries in the operand. However, the result of `PartitionGatherTrivialSlicedOperandDimensions` will do not handle these indices. Namely, if the indices is out of bound, we do not retrieve the entries from the operand and the result is 0.

This is a execution bug in SPMD partitioner in both GSPMD and Shardy. The compilation succeeds. This issue does not exist in scatter since scatter does not need to clamp the indices.

This change fixes this issue by clamping the indices at the very beginning of `PartitionGatherTrivialSlicedOperandDimensions`.

PiperOrigin-RevId: 704842079